requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
pypiwin32==223
python-dotenv==0.20.0
pywin32==304
requests==2.28.1
typing_extensions==4.2.0
urllib3==1.26.9
win10toast==0.9